removed duplicate code from test_declaration Made float to int in operator_precedence_folding improved undeclared c file improved incopatible types c file removed float from operator_precedence_folding c file added c file for a really long var for testing added c file mixing_reref_and_mult for testing added c file multiline_code for testing added c file reref_in_the_mix for testing

+++ tests/input/operator_precedence_folding.c
@@ -1,6 +1,6 @@
 // Tests for binary and unary operations combined, whether the operator precedence has been maintained
-float x = (4*2+1)-3;
-float y = (42%30)/4;
-float z = (8+5);
-float a = 1 && 5-2*2;
-float b = 3*(20+100)-291;
+int x = (4*2+1)-3;
+int y = (42%30)/4;
+int z = (8+5);
+int a = 1 && 5-2*2;
+int b = 3*(20+100)-291;
